+++ protected/humhub/modules/admin/views/setting/mobile_app.php
@@ -24,7 +24,13 @@ use yii\web\View;
 
 <?= $form->field($model, 'enableLinkService')->checkbox() ?>
 
-<?= $form->field($model, 'whiteListedDomains')->textInput() ?>
+<?= $form->field($model, 'whiteListedUrls')->widget(\kartik\widgets\Select2::class, [
+    'options' => ['multiple' => true],
+    'pluginOptions' => [
+        'tags' => true,
+        'tokenSeparators' => [','],
+    ],
+]) ?>
 
 <?= $form->beginCollapsibleFields(Yii::t('AdminModule.settings', 'Well-known files')) ?>
 <div class="form-text"><?= Yii::t('AdminModule.settings', 'Allow establishing verified connections with the mobile app to enable Android app links and iOS universal links and redirect web content to the mobile app.') ?></div>